+++ tests/testthat/test-create_rmd.R
@@ -8,7 +8,6 @@ test_that("NULL arguments return error", {
 test_that("create_rmd produces correct Rmd", {
 
   skip_on_cran()
-  # skip_on_ci()
 
   pkgs <- get_pkgs_info(pkgs = "grateful", out.dir = tempdir())
   rmd <- create_rmd(pkgs.df = pkgs, out.dir = tempdir(), out.format = "Rmd")
@@ -27,7 +26,7 @@ test_that("create_rmd produces correct Rmd", {
                  "",
                  "|Package  |Version |Citation  |",
                  "|:--------|:-------|:---------|",
-                 paste0("|grateful |", utils::packageVersion("grateful"),"  |@grateful |"),
+                 paste0("|grateful |", utils::packageVersion("grateful"),"   |@grateful |"),
                  "",
                  "**You can paste this paragraph directly in your report:**",
                  "",
@@ -54,7 +53,6 @@ test_that("csl is downloaded if needed", {
 
 test_that("rendered report is produced", {
   skip_on_cran()
-  skip_on_ci()
   pkgs <- get_pkgs_info(pkgs = "grateful", out.dir = tempdir())
   rmd <- create_rmd(pkgs.df = pkgs, out.dir = tempdir(), out.format = "html")
   expect_true(file.exists(file.path(tempdir(), "grateful-report.html")))