+++ tests/testthat/test-render_citations.R
@@ -8,7 +8,6 @@ test_that("NULL arguments return error", {
 test_that("render_citations returns a report", {
 
   skip_on_cran()
-  skip_on_ci()
 
   pkgs <- get_pkgs_info(pkgs = "grateful", out.dir = tempdir())
   rmd <- create_rmd(pkgs.df = pkgs, out.dir = tempdir(), out.format = "Rmd")
@@ -22,11 +21,6 @@ test_that("render_citations returns a report", {
   rendcit <- render_citations(Rmd.file = rmd, out.dir = tempdir(), out.format = "docx")
   expect_equal(rendcit, file.path(tempdir(), "grateful-report.docx"))
   expect_true(file.exists(file.path(tempdir(), "grateful-report.docx")))
-
-  #pdf
-  rendcit <- render_citations(Rmd.file = rmd, out.dir = tempdir(), out.format = "pdf")
-  expect_equal(rendcit, file.path(tempdir(), "grateful-report.pdf"))
-  expect_true(file.exists(file.path(tempdir(), "grateful-report.pdf")))
 
   #md
   rendcit <- render_citations(Rmd.file = rmd, out.dir = tempdir(), out.format = "md")
@@ -42,5 +36,13 @@ test_that("render_citations returns a report", {
   rendcit <- render_citations(Rmd.file = rmd, out.dir = tempdir(), out.format = "tex-document")
   expect_equal(rendcit, file.path(tempdir(), "grateful-report.tex"))
   expect_true(file.exists(file.path(tempdir(), "grateful-report.tex")))
+
+
+  skip_on_ci()
+
+  #pdf
+  rendcit <- render_citations(Rmd.file = rmd, out.dir = tempdir(), out.format = "pdf")
+  expect_equal(rendcit, file.path(tempdir(), "grateful-report.pdf"))
+  expect_true(file.exists(file.path(tempdir(), "grateful-report.pdf")))
 
 })